Finished first iteration of abstract path finding

+++ src/main/java/edu/whimc/indicator/api/search/Locatable.java
@@ -2,7 +2,7 @@ package edu.whimc.indicator.api.search;
 
 public interface Locatable<T extends Locatable<T, D>, D> {
 
-  int distanceTo(T other);
+  double distanceTo(T other);
 
   D getDomain();
